requirements.txt
streamlit
pandas
requests
googletrans==4.0.0-rc1
python-docx
openpyxl
python-pptx
PyPDF2